6: Equally weight alternative cards
Before it was just randomly picking a card from all enabled cards,
which meant that a character with multiple alternates would show up
more often than one without any.  Now we randomize on a list of
unique characters, and randomly pick one of the alternates if that
character is picked.